requirements.txt
asgiref==3.2.10
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
Django==2.2.6
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
psycopg2-binary==2.8.5
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
